The new runtime testing
Adds a new enum test option in the build for testing different parts of the OS. Pass -Dtest-type=ALL_RUNTIME to run all the runtime tests. To add a new test, add the name of the test to the enum in the build.zig file and in the rt-test.py, add a class for the test and implement run_test(). Can use the exiting once for examples.