Boot 64 bit using Limine
Removed unused CpuState


Fixed memory size and alignments

Align the .text.boot section in the ELF to the page boundary as the KernelAndModules memory type assumes sections are aligned to 4096.

Any holes in the memory map provided by the bootloader are assumed unused and are reserved in the PMM.
Update to master Zig

Swap the startFn and context around.