Added MBR implementation
This adds the a MBR partition implementation. 
Refactored some file system code to have a common FS helper functions
This also enables the using the MBR and FAT32 to create the boot image removing the echfs dependency.
Moved file system boot sector to common x86 arch.
As this boot sector is written in x86, where would need to be an ARM version, so would be arch dependent.
Added MAP file path
Fixed running the x86_64 port as this wasn't making the boot image, just the elf